requirements.txt
numpy
scipy
pandas
matplotlib
tqdm
torch==2.4.0
scanpy
accelerate
scikit-learn
requests
urllib3
POT
wandb
einops
pyyaml
ott-jax
jax
scvi-tools
IPython
ema-pytorch
